Add the GPL-3.0-only license
This repository didn't have a license before, but it's a good practice
to have that and I believe GPLv3 is the right one because I like the
strong commitment to remaining open source. I'm not a lawyer, but I
don't see a difference in license texts for GPL-3.0-only and
GPL-3.0-or-later on SPDX. I chose GPL-3.0-only because I don't know what
any future version may bring so this is conservately chosen.